Fix IAA hardware not working in VM
The QPL uses IAA hardware device according to system NUMA configuration,if
the device is used in the VM (Virtual Machine) and the VM has no NUMA
configuration, the device NUMA ID is -1 that leads the device not to
work.

So if the device NUMA ID is -1, QPL no longer check the NUMA information
of the device and uses this device directly.

Signed-off-by: yliu80 <yuan1.liu@intel.com>